% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectortiles.R
\name{addVectorTiles}
\alias{addVectorTiles}
\title{addVectorTiles}
\usage{
addVectorTiles(
  map,
  server,
  layer,
  filter = "",
  layerId,
  style = list(fill = TRUE, fillColor = "purple", fillOpacity = 0.2, color = "purple",
    opacity = 0.7, weight = 0.1, radius = 0.8),
  styleHighlight = list(fill = TRUE, fillColor = "yellow", fillOpacity = 0.2, color =
    "red", opacity = 0.7, weight = 0.1, radius = 0.8),
  group = "",
  pane = "overlayPane",
  attribution = NULL,
  debug = FALSE
)
}
\arguments{
\item{map}{map object from \code{leaflet::leaflet()}}

\item{server}{URL to tile server}

\item{layer}{layer name}

\item{filter}{filter string using \href{https://github.com/CrunchyData/pg_tileserv/blob/master/hugo/content/usage/cql.md}{Common Query Language (CQL)}
syntax. Defaults to an empty string \code{""}.}

\item{layerId}{column name to uniquely identify each feature in layer}

\item{style}{style as a list object}

\item{styleHighlight}{style as a list object for clicked feature}

\item{group}{group}

\item{pane}{pane in leaflet map}

\item{attribution}{attribution for layer}

\item{debug}{show the JavaScript code for the \code{map |> onRender(js_lyr)}}
}
\value{
map object from \code{leaflet::leaflet()} with vector tile layer added
}
\description{
Add vector tiles layer to leaflet map using the \href{https://leaflet.github.io/Leaflet.VectorGrid/vectorgrid-api-docs.html}{Leaflet.VectorGrid API reference}.
}
\examples{
library(leaflet)

leaflet() |>
  addTiles() |>
  setView(-122.4, 33.8, 4) |>
  addVectorTiles(
    server  = "https://tile.calcofi.io",
    layer   = "public.stations",
    layerId = "stationid")
}
\concept{vectortiles}
